+++ esp-hal/src/i2c/mod.rs
@@ -37,10 +37,10 @@
 //!
 //! // Create a new peripheral object with the described wiring
 //! // and standard I2C clock speed.
-//! let mut i2c = I2c::new(
-//!     peripherals.I2C0,
-//!     100.kHz(),
-//! )
+//! let mut i2c = I2c::new_with_config(peripherals.I2C0, Config {
+//!     frequency: 100.kHz(),
+//!     timeout: None,
+//! })
 //! .with_sda(io.pins.gpio1)
 //! .with_scl(io.pins.gpio2);
 //!
@@ -55,8 +55,9 @@
 mod support;
 mod version;
 
-use core::marker::PhantomData;
+use core::{convert::Infallible, marker::PhantomData};
 
+use embassy_embedded_hal::SetConfig;
 use embassy_sync::waitqueue::AtomicWaker;
 use fugit::HertzU32;
 use version::{I2C_CHUNK_SIZE, I2C_LL_INTR_MASK};
@@ -205,12 +206,44 @@ impl From<Ack> for u32 {
     }
 }
 
+/// I2C driver configuration
+#[derive(Debug, Clone, Copy)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
+pub struct Config {
+    /// The I2C clock frequency.
+    pub frequency: HertzU32,
+
+    /// The I2C timeout.
+    // TODO: explain this function better - what's the unit, what happens on
+    // timeout, and just what exactly is a timeout in this context?
+    pub timeout: Option<u32>,
+}
+
+impl Default for Config {
+    fn default() -> Self {
+        use fugit::RateExtU32;
+        Config {
+            frequency: 100.kHz(),
+            timeout: None,
+        }
+    }
+}
+
 /// I2C driver
 pub struct I2c<'d, DM: Mode, T = AnyI2c> {
     i2c: PeripheralRef<'d, T>,
     phantom: PhantomData<DM>,
-    frequency: HertzU32,
-    timeout: Option<u32>,
+    config: Config,
+}
+
+impl<T: Instance, DM: Mode> SetConfig for I2c<'_, DM, T> {
+    type Config = Config;
+    type ConfigError = Infallible;
+
+    fn set_config(&mut self, config: &Self::Config) -> Result<(), Self::ConfigError> {
+        self.apply_config(config);
+        Ok(())
+    }
 }
 
 impl<'d, T, DM: Mode> I2c<'d, DM, T>
@@ -228,16 +261,13 @@ where
         PeripheralClockControl::reset(self.i2c.peripheral());
         PeripheralClockControl::enable(self.i2c.peripheral());
 
-        self.driver().setup(self.frequency, self.timeout);
+        self.driver().setup(&self.config);
     }
 
-    /// Set the I2C timeout.
-    // TODO: explain this function better - what's the unit, what happens on
-    // timeout, and just what exactly is a timeout in this context?
-    pub fn with_timeout(mut self, timeout: Option<u32>) -> Self {
-        self.timeout = timeout;
-        self.driver().setup(self.frequency, self.timeout);
-        self
+    /// Applies a new configuration.
+    pub fn apply_config(&mut self, config: &Config) {
+        self.config = *config;
+        self.driver().setup(&self.config);
     }
 
     fn transaction_impl<'a>(
@@ -335,11 +365,20 @@ where
 }
 
 impl<'d> I2c<'d, Blocking> {
-    /// Create a new I2C instance
+    /// Creates a new I2C instance.
+    ///
     /// This will enable the peripheral but the peripheral won't get
     /// automatically disabled when this gets dropped.
-    pub fn new(i2c: impl Peripheral<P = impl Instance> + 'd, frequency: HertzU32) -> Self {
-        Self::new_typed(i2c.map_into(), frequency)
+    pub fn new(i2c: impl Peripheral<P = impl Instance> + 'd) -> Self {
+        Self::new_with_config(i2c.map_into(), Config::default())
+    }
+
+    /// Creates a new I2C instance with a given configuration.
+    ///
+    /// This will enable the peripheral but the peripheral won't get
+    /// automatically disabled when this gets dropped.
+    pub fn new_with_config(i2c: impl Peripheral<P = impl Instance> + 'd, config: Config) -> Self {
+        Self::new_typed_with_config(i2c.map_into(), config)
     }
 }
 
@@ -347,23 +386,31 @@ impl<'d, T> I2c<'d, Blocking, T>
 where
     T: Instance,
 {
-    /// Create a new I2C instance
+    /// Creates a new I2C instance with a given configuration.
+    ///
     /// This will enable the peripheral but the peripheral won't get
     /// automatically disabled when this gets dropped.
-    pub fn new_typed(i2c: impl Peripheral<P = T> + 'd, frequency: HertzU32) -> Self {
+    pub fn new_typed(i2c: impl Peripheral<P = T> + 'd) -> Self {
+        Self::new_typed_with_config(i2c, Config::default())
+    }
+
+    /// Creates a new I2C instance with a given configuration.
+    ///
+    /// This will enable the peripheral but the peripheral won't get
+    /// automatically disabled when this gets dropped.
+    pub fn new_typed_with_config(i2c: impl Peripheral<P = T> + 'd, config: Config) -> Self {
         crate::into_ref!(i2c);
 
         let i2c = I2c {
             i2c,
             phantom: PhantomData,
-            frequency,
-            timeout: None,
+            config,
         };
 
         PeripheralClockControl::reset(i2c.i2c.peripheral());
         PeripheralClockControl::enable(i2c.i2c.peripheral());
 
-        i2c.driver().setup(frequency, None);
+        i2c.driver().setup(&i2c.config);
         i2c
     }
 
@@ -376,8 +423,7 @@ where
         I2c {
             i2c: self.i2c,
             phantom: PhantomData,
-            frequency: self.frequency,
-            timeout: self.timeout,
+            config: self.config,
         }
     }
 
@@ -531,8 +577,7 @@ where
         I2c {
             i2c: self.i2c,
             phantom: PhantomData,
-            frequency: self.frequency,
-            timeout: self.timeout,
+            config: self.config,
         }
     }
 
@@ -866,7 +911,7 @@ struct Driver<'a> {
 impl Driver<'_> {
     /// Configures the I2C peripheral with the specified frequency, clocks, and
     /// optional timeout.
-    fn setup(&self, frequency: HertzU32, timeout: Option<u32>) {
+    fn setup(&self, config: &Config) {
         self.info.register_block().ctr().write(|w| {
             // Set I2C controller to master mode
             w.ms_mode().set_bit();
@@ -887,7 +932,7 @@ impl Driver<'_> {
         self.set_filter(Some(7), Some(7));
 
         // Configure frequency
-        self.set_frequency(frequency, timeout);
+        self.set_frequency(config.frequency, config.timeout);
 
         self.update_config();