+++ examples/src/bin/i2c_bmp180_calibration_data.rs
@@ -23,7 +23,7 @@ fn main() -> ! {
 
     // Create a new peripheral object with the described wiring and standard
     // I2C clock speed:
-    let mut i2c = I2c::new(peripherals.I2C0, 100.kHz())
+    let mut i2c = I2c::new(peripherals.I2C0)
         .with_sda(io.pins.gpio4)
         .with_scl(io.pins.gpio5);