+++ hil-test/tests/i2c.rs
@@ -8,7 +8,6 @@
 use esp_hal::{
     gpio::Io,
     i2c::{I2c, Operation},
-    prelude::*,
     Async,
     Blocking,
 };
@@ -40,11 +39,9 @@ mod tests {
 
         // Create a new peripheral object with the described wiring and standard
         // I2C clock speed:
-        let i2c = I2c::new(peripherals.I2C0, 100.kHz())
-            .with_sda(sda)
-            .with_scl(scl);
-
-        Context { i2c }
+        Context {
+            i2c: I2c::new(peripherals.I2C0).with_sda(sda).with_scl(scl),
+        }
     }
 
     #[test]